% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{add_filter_ptm_pos_rowdata_mq}
\alias{add_filter_ptm_pos_rowdata_mq}
\title{Add rowData columns with positions of PTMs with respect to peptide sequence}
\usage{
add_filter_ptm_pos_rowdata_mq(
  obj,
  ptms_to_retain = c("Phospho (STY)"),
  prob_col = "Phospho..STY..Probabilities",
  min_prob = 0.501,
  filter_pep_by_prob = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment}. Proteomics dataset}

\item{ptms_to_retain}{\code{character} vector with PTMs to retain. Inspect values in rowData(obj)$Modified.sequence to get correct PTM name}

\item{prob_col}{\code{character} name of column containing PTM probabilities}

\item{min_prob}{\code{numeric} Minimum acceptable probability for PTM localisation}

\item{filter_pep_by_prob}{\code{logical} Filter the output to only return cases where the number of sites passing the probability threshold
equals the number of PTMs in the peptide}

\item{verbose}{\code{logical} Describe the number of PTMs detected}
}
\value{
Returns a \code{SummarizedExperiment} with an additional column in the RowData describing the position of the PTMs with respect to the peptide sequence
}
\description{
This function takes a \code{SummarizedExperiment} object with the evidence.txt
output from MaxQuant and adds rowData columns to describe the PTMs present in the peptide and their positions within the peptide' '
}
