% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_protein_name.R
\name{add_gene_long_protein_name_pd}
\alias{add_gene_long_protein_name_pd}
\title{Extract the gene name and long-form protein name from the master protein descriptions column}
\usage{
add_gene_long_protein_name_pd(
  obj,
  master_protein_desc_col = "Master.Protein.Descriptions",
  gene_name_out_col = "Master.Protein.gene.name",
  long_protein_name_out_col = "Master.Protein.long.name"
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer}

\item{master_protein_desc_col}{\code{string}. Name of column containing master
proteins descriptions.}

\item{gene_name_out_col}{\code{string}. Name of output column containing the gene names}

\item{long_protein_name_out_col}{\code{string}. Name of output column containing the long format
protein names}
}
\value{
Returns a \code{QFeatures} with the filtered Proteome Discoverer output.
}
\description{
This function extracts the gene name and long-form protein name
from the master protein descriptions column in output from
Proteome Discoverer for DDA.

It assumes the master protein description column is in the format
(.\emph{) .} GN=(.\emph{) PE.}, where the first group is the long format of the protein
name and the second group is the gene name. Where the gene name is not included,
an empty string is returned
An example of the expected input in the column is:
Serum albumin OS=Bos taurus GN=ALB PE=1 SV=4

\if{html}{\out{<div class="sourceCode">}}\preformatted{}\if{html}{\out{</div>}}
}
