% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{add_independent_filtering_padj}
\alias{add_independent_filtering_padj}
\title{Perform independent filtering to threshold the over-representation testing
on the number of features in each category to limit the multiple testing burden}
\usage{
add_independent_filtering_padj(
  obj,
  alpha = 0.1,
  theta = seq(0, 1, 0.01),
  p_value_col = "over_represented_pvalue",
  filter_col = "numInCat",
  plot = TRUE
)
}
\arguments{
\item{obj}{\code{data.frame} containing \code{goseq} results as generated by
\code{\link{get_enriched_go}}
See below for an example.}

\item{alpha}{\code{numeric} alpha value to use for rejection of null hypothesis.
Note this is only used for optimising the threshold and any alternative value for alpha can be used downstream}

\item{theta}{\code{numeric} vector of thresholds (fractions of the data to remove)}

\item{p_value_col}{\code{character} Column with p-values}

\item{filter_col}{\code{character} Column with filtering criteria values}

\item{plot}{\code{logical} Plot the relationship between the}
}
\value{
Returns the \code{data.frame} with an added column denoting the optimised adjusted p-value ('padj_if')
}
\description{
When testing for enrichment of feature annotations, we frequently test
thousands of annotations simultaneously, which increases the multiple testing burden considerably.
Given many annotations have very few annotated features, there will be too little power to detect
over-representation. Inspired by DESeq2, which was in turn inspired by \href{https://www.pnas.org/doi/10.1073/pnas.0914005107}{Bourgon, Gentleman, and Huber (2010)},
this function applies the principle of 'independent filtering' to threshold the minimum
number of features per annotation (in the background) to optimally limit the multiple testing burden. This is
statistically valid because the filtering criteria (number of features per annotation) is
statistically independent from the test statistic under the null hypothesis.
The independent filtering works because the filtering criteria is correlated
with the test statistic under the alternative hypothesis. For further justification and discussion of independent filtering,
see \href{https://www.pnas.org/doi/10.1073/pnas.0914005107}{Bourgon, Gentleman, and Huber (2010)} and
the \href{https://bioconductor.org/packages/devel/bioc/vignettes/DESeq2/inst/doc/DESeq2.html#indfilttheory}{DESeq2 vignette}.

The code function is largely lifted from \code{DEseq:::pvalueAdjustment}

Please cite \href{https://www.pnas.org/doi/10.1073/pnas.0914005107}{Bourgon, Gentleman, and Huber (2010)} when using this function.
}
