% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_position.R
\name{add_peptide_positions}
\alias{add_peptide_positions}
\title{Add a column describing the position of the peptide sequence with respect
to the protein}
\usage{
add_peptide_positions(
  obj,
  proteome_fasta,
  master_protein_col = "Master.Protein.Accessions"
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment} with PD output at PSM/peptide level}

\item{proteome_fasta}{\code{character} Filepath for proteome fasta}

\item{master_protein_col}{\code{character} Column name for master protein}
}
\value{
\code{SummarizedExperiment} with extra row columns detailing the peptide position
}
\description{
Identify the position of the peptide sequence in the protein.

The peptide position is undefined (NA) if:
\enumerate{
\item The peptide sequence has multiple master proteins
\item The protein does not exist in the proteome fasta
\item The peptide is repeated in the protein sequence.
}

The peptide_position_info column details why the peptide positions columns are NA
}
