% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{add_peptide_positions_from_cleavage}
\alias{add_peptide_positions_from_cleavage}
\title{Add peptide positions with respect to the protein}
\usage{
add_peptide_positions_from_cleavage(
  obj,
  proteome_fasta,
  digest_enzyme = "trypsin-simple",
  missed_cleavages = c(0, 1, 2),
  master_protein_col = "Leading.razor.protein",
  sequence_col = "Sequence",
  start_col = "start",
  end_col = "end"
)
}
\arguments{
\item{obj}{SummarizedExperiment. Proteomics dataset}

\item{proteome_fasta}{character filepath to fasta with protein sequences}

\item{digest_enzyme}{character. Enzyme used. See ?cleaver::cleave}

\item{missed_cleavages}{numeric. Vector of allowed number of missed cleavages}

\item{master_protein_col}{character. Name of column containing master proteins}

\item{sequence_col}{character. Name of column containing peptide sequences}

\item{start_col}{character. Name of output column containing peptide start positions}

\item{end_col}{character. Name of output column containing peptide end positions}
}
\value{
Returns a SummarizedExperiment with additional rowData columns describing
peptide start and end positions with respect to the protein
}
\description{
This function takes a SummarizedExperiment object and adds rowData
columns to describe the positions of the peptides within the proteins', taking
into account the possible peptides according to the digestion enzyme used.
}
