% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{add_ptm_pos_rowdata_mq}
\alias{add_ptm_pos_rowdata_mq}
\title{Add rowData columns with positions of PTMs with respect to peptide sequence}
\usage{
add_ptm_pos_rowdata_mq(
  obj,
  ptms_to_retain = c("p", "(ph)", "(ac)p"),
  ptm_encoding_pos = c(p = -1, `(ph)` = 1, `(ac)p` = -1, `(ox)` = 1),
  verbose = TRUE
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment}. Proteomics dataset}

\item{ptms_to_retain}{\code{character} vector with PTMs to retain. Inspect values in rowData(obj)$Modified.sequence to get correct PTM name}

\item{ptm_encoding_pos}{\code{character} name vector describing whether modifification comes before (-1) or after (1) amino acid in Modified.sequence column values.}

\item{verbose}{\code{logical} Describe the number of PTMs detected}
}
\value{
Returns a \code{SummarizedExperiment} with an additional column in the RowData describing the position of the PTMs with respect to the peptide sequence
}
\description{
This function takes a \code{SummarizedExperiment} object with the evidence.txt
output from MaxQuant and adds rowData columns to describe the PTMs present in the peptide and their positions within the peptide' '
}
