% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{add_site_sequence}
\alias{add_site_sequence}
\title{Add a column with amino acid sequence around a PTM}
\usage{
add_site_sequence(
  obj,
  proteome_fasta,
  master_protein_col = "Master.Protein.Accessions",
  sequence_pad = 7
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment}. Proteomics dataset}

\item{proteome_fasta}{\code{character} Filepath for proteome fasta}

\item{master_protein_col}{\code{character} Column name for master protein}

\item{sequence_pad}{\code{numeric} Number of amino acids to include on either side of PTM}
}
\value{
\code{SummarizedExperiment}
}
\description{
Add a rowData column with amino acid sequence around a PTM. Value will be
NA if peptide maps to multiple proteins or has multiple PTMs. If padding extends
outside the protein AA sequence, padding will be extended with '_'. The PTM-
centered AA sequence is useful to integrate external databases. Input
should have been been passed through \code{add_ptm_positions}
to add the 'ptm_positions_prot' column to the rowData.
}
