% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_position.R
\name{add_tryptic_termini_flags}
\alias{add_tryptic_termini_flags}
\title{Evaluate peptide termini against theoretical cleavage positions}
\usage{
add_tryptic_termini_flags(
  obj,
  proteome_fasta,
  digest_enzyme = "trypsin-simple",
  missed_cleavages = c(0, 1, 2),
  master_protein_col = "Leading.razor.protein",
  start_col = "start",
  end_col = "end"
)
}
\arguments{
\item{obj}{SummarizedExperiment}

\item{proteome_fasta}{character. Path to FASTA file}

\item{digest_enzyme}{character. Enzyme for cleaver::cleavageRanges}

\item{missed_cleavages}{numeric vector. Allowed missed cleavages}

\item{master_protein_col}{character. Column in rowData containing protein IDs}

\item{start_col}{character. Column containing peptide start positions}

\item{end_col}{character. Column containing peptide end positions}
}
\value{
SummarizedExperiment with additional rowData columns:
N_tryp and C_tryp
}
\description{
Evaluate peptide termini against theoretical cleavage positions
}
