% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{center_normalise_to_ref}
\alias{center_normalise_to_ref}
\title{Center-median normalise the expression matrix in an MSnSet using medians
from a reference dataset}
\usage{
center_normalise_to_ref(
  obj,
  medians,
  center_to_zero = FALSE,
  on_log_scale = FALSE
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment}}

\item{medians}{\verb{vector, numeric}. Sample medians from reference dataset}

\item{center_to_zero}{\code{logical}. Centre the data range on zero.
If FALSE, normalisation retains original data range.}

\item{on_log_scale}{\code{logical}. Input data is log-transformed}
}
\value{
Returns a \code{SummarisedExperiment} with the assay data column center-median
normalised
}
\description{
Center-median normalisation is a simple normalisation method
that is appropriate for relative abundance proteomics such as isobaric tagging.
This can be achieved with \code{QFeatures::normalize(method='diff.median')}. However,
for some experimental designs, the normalisation should be against the medians
in another dataset. For example, for PTM studies, one may wish to isobaric
tag samples, pool, and then PTM-enriched, with the enriched sample quantified
in a separate run to the non-enriched (total) sample. In this case, it may make
more sense to center-median normalise the PTM-enriched samples using the median
from the total samaples.
}
\examples{
# normalise one assay using medians taken from another, as in a PTM
# experiment where the total fraction supplies the reference
reference_medians <- get_medians(tmt_qf[["protein"]])

normalised <- center_normalise_to_ref(
  tmt_qf[["protein"]], reference_medians, on_log_scale = TRUE)

round(get_medians(normalised), 3)
}
