% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fasta.R
\name{check_uniprot_release}
\alias{check_uniprot_release}
\title{Check the current UniProt release}
\usage{
check_uniprot_release()
}
\value{
Returns a \code{character}, the current release number in the format
YYYY_XX where YYYY is the calendar year and XX a 2-digit number that
is incremented for each release of a given year, e.g. 2010_01, 2010_02, etc.
}
\description{
UniProt releases are published approximately every 8 weeks.
This function checks what the current UniProt release is.
}
\examples{
\dontrun{
# print release number to console
check_uniprot_release()

# save release number and use in e.g. a file name
rls <- check_uniprot_release()

paste0("folder/filename_", rls, ".fasta")
}

}
