% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{create_long_form_imputed_data}
\alias{create_long_form_imputed_data}
\title{Create long format data with column defining if quantification value is imputed}
\usage{
create_long_form_imputed_data(
  obj,
  i_unimputed,
  i_imputed,
  column_variables,
  row_variables
)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i_unimputed}{\code{string}. Index for the SummarizedExperiment with data without imputation}

\item{i_imputed}{\code{string}. Index for the SummarizedExperiment with data with imputation}

\item{column_variables}{\verb{character vector}. Variables in \code{colData(obj)} to include}

\item{row_variables}{\verb{character vector}. Variables in \code{rowData(obj)} to include}
}
\value{
Returns a \code{data.frame} with quantification data in long form and including
a \code{is.imputed} column
}
\description{
This function takes a \code{Qfeatures} object with unimputed data and
imputed data in separate assays and creates a long format data.frame with a column
called \code{is.imputed} which describes whether the quantification value was imputed
}
