% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw.R
\name{filter_features}
\alias{filter_features}
\title{Filter Proteome Discoverer output}
\usage{
filter_features(
  obj,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  silac = FALSE,
  TMT = FALSE,
  level = "peptide",
  filter_crap = TRUE,
  crap_proteins = NULL,
  filter_associated_crap = TRUE,
  remove_no_quant = TRUE
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer.
Use \code{\link[QFeatures]{readQFeatures}} to read in .txt file}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{protein_col}{\code{string}. Name of column containing all protein
matches.}

\item{unique_master}{\code{logical}. Filter out features without a unique
master protein.}

\item{silac}{\code{logical}. Is the experiment a SILAC experiment?}

\item{TMT}{\code{logical}. Is the experiment a TMT experiment?}

\item{level}{\code{string}. Type of input file, must be one of either
\code{"peptide"} or \code{"PSM"}.}

\item{filter_crap}{\code{logical}. Filter out features which match a cRAP
protein.}

\item{crap_proteins}{\verb{character vector}. The protein IDs form the cRAP proteins,}

\item{filter_associated_crap}{\code{logical}. Filter out features which
match a cRAP associated protein.}

\item{remove_no_quant}{\code{logical}. Remove features with no quantification}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the output .txt files (peptide groups or PSMs) from
Proteome Discoverer based on various criteria.
\enumerate{
\item Remove features without a master protein
\item (Optional) Remove features without a unique master protein  (i.e.
Number.of.Protein.Groups == 1)
\item (Optional) Remove features matching a cRAP protein
\item (Optional) Remove features matching any protein associated with
a cRAP protein (see below)
\item Remove features without quantification values (only if TMT or SILAC
are \code{TRUE} and \code{level = "peptide"}.)
}
}
\details{
\strong{Associated cRAP proteins} are proteins which have at least
one feature shared with a cRAP protein. It has been observed that the cRAP
fasta files often do not contain all possible cRAP proteins e.g. some features
can be assigned to a keratin which is not in the provided cRAP database.

In the example below, using \code{filter_associated_crap = TRUE} will filter out f2 and f3 in
addition to f1, regardless of the value in the Master.Protein.Accession column.

\if{html}{\out{<div class="sourceCode">}}\preformatted{feature  Protein.Accessions         Master.Protein.Accessions
f1       protein1, protein2, cRAP,  protein1,
f2       protein1, protein3         protein3,
f3       protein2                   protein2
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{

#### PSMs.txt example ####
# load PD PSMs.txt output
tmt_qf <- readQFeatures(assayData = psm_tmt_total,
 quantCols = 36:45,
 name = "psms_raw")

# extract the UniProt accessions from the cRAP FASTA headers
crap_accessions <- get_crap_fasta_accessions(crap_fasta_inf)

# filter thg peptides
psm2 <- parse_features(
  obj = tmt_qf[['psms_raw']],
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  TMT = TRUE,
  level = "PSM",
  filter_crap = TRUE,
  crap_proteins = crap_accessions,
  filter_associated_crap = TRUE
)


}
}
