% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw.R
\name{filter_features_pd_dda}
\alias{filter_features_pd_dda}
\title{Filter Proteome Discoverer DDA output}
\usage{
filter_features_pd_dda(
  obj,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  proteotypic = FALSE,
  filter_contaminant = TRUE,
  contaminant_proteins = NULL,
  crap_proteins = NULL,
  filter_associated_contaminant = TRUE,
  remove_no_quant = TRUE,
  cont_string = "Cont_"
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer.
Use \code{\link[QFeatures]{readQFeatures}} to read in .txt file}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{protein_col}{\code{string}. Name of column containing all protein
matches.}

\item{unique_master}{\code{logical}. Filter out features where the master protein
column does not resolve to a single protein accession.}

\item{proteotypic}{\code{logical}. Filter out features whose peptide sequence is
found in more than one protein.}

\item{filter_contaminant}{\code{logical}. Filter out features which match a contaminant
protein.}

\item{contaminant_proteins}{\verb{character vector}. The protein IDs form the contaminant proteins}

\item{crap_proteins}{\verb{character vector}. Same as contaminant_proteins. Available for backwards compatibility. Default is NULL. If both contaminant_proteins and crap_proteins are set, an error is thrown.}

\item{filter_associated_contaminant}{\code{logical}. Filter out features which
match a contaminant associated protein.}

\item{remove_no_quant}{\code{logical}. Remove features with no quantification}

\item{cont_string}{\code{string}. string to search for contaminants}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the output .txt files (peptide groups or PSMs) from
Proteome Discoverer for DDA, based on various criteria:
\enumerate{
\item Remove features without a master protein
\item Remove features without a unique master protein  (i.e.
Number.of.Protein.Groups == 1)
\item Remove features which are not proteotypic (i.e. Number.of.Proteins == 1)
\item Remove features matching a contaminant protein
\item Remove features matching any protein associated with
a contaminant protein (see below)
\item Remove features without quantification values
}
}
\details{
\code{unique_master} and \code{proteotypic} are different filters.
\code{unique_master} asks whether Proteome Discoverer resolved the feature to a
single protein accession, so it removes features which are ambiguous between
protein groups. \code{proteotypic} asks whether the peptide sequence occurs in
only one protein in the database, so it also removes features whose protein
group holds several indistinguishable proteins, even where a master was
assigned. See \code{vignette("gotcha_peptide_to_protein")}.

\strong{Associated contaminant proteins} are proteins which have at least
one feature shared with a contaminant protein. It has been observed that the contaminant
fasta files often do not contain all possible contaminant proteins e.g. some features
can be assigned to a keratin which is not in the provided contaminant database.

In the example below, using \code{filter_associated_contaminant = TRUE} will filter out f2 and f3 in
addition to f1, regardless of the value in the Master.Protein.Accession column.

\if{html}{\out{<div class="sourceCode">}}\preformatted{feature  Protein.Accessions         Master.Protein.Accessions
f1       protein1, protein2, contaminant,  protein1,
f2       protein1, protein3         protein3,
f3       protein2                   protein2
}\if{html}{\out{</div>}}
}
\examples{
# load PD PSM-level output
tmt_qf <- QFeatures::readQFeatures(assayData = psm_tmt_clock,
  colData = tmt_clock_design,
  quantCols = rownames(tmt_clock_design),
  name = "psms_raw")

# extract the accessions from the contaminant FASTA, in both the prefixed
# and bare forms, since the search may not have renamed its entries
contaminant_fasta <- system.file(
  "extdata", "0602_Universal_Contaminants.fasta.gz", package = "biomasslmb")

contaminant_accessions <- get_contaminant_fasta_accessions(contaminant_fasta)
contaminant_accessions <- c(contaminant_accessions,
                            sub("^Cont_", "", contaminant_accessions))

# remove contaminants and PSMs without a unique master protein
psms_filtered <- filter_features_pd_dda(
  obj = tmt_qf[["psms_raw"]],
  contaminant_proteins = contaminant_accessions,
  filter_contaminant = TRUE,
  filter_associated_contaminant = TRUE,
  unique_master = TRUE)

c(before = nrow(tmt_qf[["psms_raw"]]), after = nrow(psms_filtered))
}
