% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_features_per_protein.R
\name{filter_features_per_protein}
\alias{filter_features_per_protein}
\title{Remove features which are assigned to a protein with too few supporting
features in total}
\usage{
filter_features_per_protein(
  obj,
  min_features,
  master_protein_col = "Master.Protein.Accessions",
  plot = FALSE
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment} with PSM or peptide-level quantification}

\item{min_features}{\code{numeric} Threshold for minimum features per protein}

\item{master_protein_col}{\code{character} Column name for master protein}

\item{plot}{Set TRUE to plot histogram of features per protein per sample}
}
\value{
\code{SummarizedExperiment}
}
\description{
For summarisation of PSM or peptide to protein, we need a
minimum number of finite values per protein per sample.
}
\examples{
# keep only features from proteins with at least 2 features
psms <- tmt_qf[["psms_filtered_sn"]]

c(before = nrow(psms),
  after = nrow(filter_features_per_protein(psms, min_features = 2)))
}
