% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{filter_maxquant_ptm}
\alias{filter_maxquant_ptm}
\title{Filter PTM data from MaxQuant to retain those with a given PTM}
\usage{
filter_maxquant_ptm(obj, i, ptm_prob_col = "Phospho..STY..Probabilities")
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarizedExperiment with data without imputation}

\item{ptm_prob_col}{\code{character} column name for PTMs.}
}
\value{
Returns a \code{SummarizedExperiment} where the value for the PTM probability column is not empty
}
\description{
This function takes a \code{Qfeatures} object with the evidence.txt
output from MaxQuant and filters it to retain peptides with a given PTM.
}
