% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_TMT_PSMs.R
\name{filter_TMT_PSMs}
\alias{filter_TMT_PSMs}
\title{Filter a PSM-level summarizedExperiment to remove low quality PSMs}
\usage{
filter_TMT_PSMs(
  obj,
  inter_thresh = 100,
  sn_thresh = 0,
  spsmm_thresh = 0,
  master_protein_col = "Master.Protein.Accessions",
  inter_col = "Isolation.Interference.in.Percent",
  sn_col = "Average.Reporter.SN",
  spsmm_col = "SPS.Mass.Matches.in.Percent",
  from_PD = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{obj}{\code{summarizedExperiment}. Should contain PSMs-level TMT quantification}

\item{inter_thresh}{\code{numeric}. Maximum allowed interference/co-isolation}

\item{sn_thresh}{\code{numeric}. Minimum allowed average signal:}

\item{spsmm_thresh}{\code{numeric}. Minimum allowed value for Synchronous precursor selection mass matches (\%) - SPS-MM, e.g the percentage of MS3 fragments that can be explicitly traced back to the precursor peptides}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{inter_col}{\code{string}. Name of column containing the interference value.}

\item{sn_col}{\code{string}. Name of column containing the signal:noise value.}

\item{spsmm_col}{\code{string}. Name of column containing the SPS-MM value.}

\item{from_PD}{\code{logical}. Input is from ProteomeDiscover. If set, will filter using
Quan.Info and PSM.Ambiguity columns. Default is TRUE}

\item{verbose}{\code{logical}. Default is TRUE, use verbose output messages.}
}
\value{
Returns an \code{summarizedExperiment} with the filtered PSMs.
}
\description{
Filter PSMs from TMT quantification to remove the following:
\enumerate{
\item Missing values (\code{NA}) for all tags
\item Interference/co-isolation above a set value (default=100, e.g no filtering)
\item Signal:noise ratio below a set value (default=0, e.g no filtering)
\item SPS-MM below a set value (default=0, e.g no filtering)
\item Quan.Info is not empty ('')
\item PSM.Ambiguity is not 'Selected' or 'Unambiguous'
}
}
\examples{
tmt_qf <- QFeatures::readQFeatures(assayData = psm_tmt_clock,
  colData = tmt_clock_design,
  quantCols = rownames(tmt_clock_design),
  name = "psms_raw")

# a more accurate average signal:noise than PD reports
tmt_qf[["psms_raw"]] <- update_average_sn(tmt_qf[["psms_raw"]])

# drop PSMs with low signal:noise or high co-isolation
psms_filtered <- filter_TMT_PSMs(tmt_qf[["psms_raw"]],
                                 inter_thresh = 50, sn_thresh = 10)

nrow(psms_filtered)
}
