% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{get_contaminant_fasta_accessions}
\alias{get_contaminant_fasta_accessions}
\title{Extract the contaminants protein accessions from a contaminants fasta file}
\usage{
get_contaminant_fasta_accessions(cont_fasta_inf)
}
\arguments{
\item{cont_fasta_inf}{\code{character} Filepath to the contaminants fasta}
}
\value{
Returns a \code{list} of contaminant IDs
}
\description{
Extract the contaminants protein accessions from a contaminants fasta file
}
