% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_go.R
\name{get_go_terms}
\alias{get_go_terms}
\title{Obtain GO term annotations for proteins}
\usage{
get_go_terms(UniprotID, expand_terms = FALSE, verbosity = 0)
}
\arguments{
\item{UniprotID}{\verb{character vector} Uniprot IDs}

\item{expand_terms}{\code{logical} Should GO terms be expanded to include all ancestors}

\item{verbosity}{\code{integer} Verbosity level for uniprotREST::uniprot_map}
}
\value{
\code{data.frame} object.
}
\description{
Given a set of UniProt IDs, this function queries UniProt to obtain
the annotated GO terms. Optionally, these GO terms can be expanded to include all ancestors too, which
can be helpful when using GO over-representation/enrichment tools that do not
consider the GO term heirarchy. Note that expansion will significantly increase the run-time
}
\examples{
\dontrun{
uniprotIDs <- c('O76024', 'Q03135', 'Q96T23')
get_go_terms(uniprotIDs, expand_terms=TRUE)
}
}
