% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.R
\name{get_maxquant_cont_accessions}
\alias{get_maxquant_cont_accessions}
\title{Extract the contaminants protein accessions from a MaxQuant contaminants fasta file}
\usage{
get_maxquant_cont_accessions(cont_fasta_inf = NULL)
}
\arguments{
\item{cont_fasta_inf}{\code{character} Filepath to the contaminants fasta. Defaults to NULL
in which case, contaminants.fasta.gz from this package will be used}
}
\value{
Returns a \code{list} of contaminant IDs
}
\description{
Extract the contaminants protein accessions from a MaxQuant contaminants fasta file
}
