% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{get_medians}
\alias{get_medians}
\title{Extract the assay column medians from an MSnSet}
\usage{
get_medians(obj)
}
\arguments{
\item{obj}{\code{SummarisedExperiment}}
}
\value{
\code{vector} of assay column medians in colnames order
}
\description{
Extract the assay column medians from an MSnSet
}
\examples{
# per-sample median abundance, used as a normalisation reference
get_medians(tmt_qf[["protein"]])
}
