% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_tm_toplogy.R
\name{get_protein_tm_topology}
\alias{get_protein_tm_topology}
\title{Obtain protein transmembrane-domains and topology information}
\usage{
get_protein_tm_topology(uniprotIDs, verbosity = 0)
}
\arguments{
\item{uniprotIDs}{\verb{character vector} Uniprot IDs}

\item{verbosity}{\code{integer} Verbosity level for uniprotREST::uniprot_map}
}
\value{
\code{data.frame} object.
}
\description{
Given a set of UniProt IDs, this function queries UniProt to obtain
the information regarding transmembrane domains (TMDs) and topology and then parses
this information to define additional columns describining the TMDs and topology
}
\examples{
\dontrun{
uniprotIDs <- c('O76024', 'Q03135', 'Q96T23')
get_protein_tm_topology(uniprotIDs)
}
}
