% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_detection.R
\name{get_samples_present}
\alias{get_samples_present}
\title{Extract the number of samples each feature was detected in for each experiment in a Qfeatures object}
\usage{
get_samples_present(obj, rowVars, rename_cols = NULL)
}
\arguments{
\item{obj}{\code{QFeatures} object}

\item{rowVars}{\verb{character vector} row variables to group by, e.g 'Master.Protein.Accessions'}

\item{rename_cols}{\verb{named character vector} optional named list to rename the experiments. List values should be current experiment names and list names should be updated experiment names}
}
\value{
\code{data.frame} object.
}
\description{
When exploring the proteome coverage, it is instructive to consider how many samples a protein
was present in for each level of filtering. This function takes a Qfeatures object and tallies how many samples a variable in the \code{rowData}
was present in. When the row variables are the protein names, the tally is thus the number of samples the protein was present in.
}
\examples{
rename_cols <- c("All PSMs" = "psms_raw", "Protein" = "protein")

samples_present <- get_samples_present(
  tmt_qf, rowVars = "Master.Protein.Accessions", rename_cols = rename_cols)

head(samples_present)
}
