% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{lfq_qf}
\alias{lfq_qf}
\title{LFQ-DDA data}
\format{
An object of class \code{QFeatures} of length 5.
}
\usage{
lfq_qf
}
\description{
Qfeature containing LFQ-DDA data processed in the \verb{LFQ-DDA QC Peptide-level quantification and summarisation to protein-level abundance} vignette
}
\keyword{datasets}
