% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_features_per_protein.R
\name{mask_protein_level_quant}
\alias{mask_protein_level_quant}
\title{Replace protein-level quantifications with NA if they derive from too few lower feature level quantifications}
\usage{
mask_protein_level_quant(obj, retain_mask)
}
\arguments{
\item{obj}{\code{SummarizedExperiment} with PSM or peptide-level quantification}

\item{retain_mask}{\code{matrix} detailing whether a protein-level quantification had sufficient lower level
quantification values for each sample. Can be obtained with \code{\link{get_protein_no_quant_mask }}}
}
\value{
\code{SummarizedExperiment} with quantification values replaced by NA where they derive from too few lower feature level quantifications
}
\description{
Protein level abundances are more accurately quantified where there are more
features (PSMs/peptides) to summarise from.

Usually, we are performing the summarisation
from a matrix (columns=samples, rows=features) with an associated feature to protein ID mapping.
Within the matrix, some values may be missing (NA). In order to correctly identify which proteins can
be quantified, we need to start from the feature level object and create a mask
which we can use to replace protein-level quantification values with NA.
This is can be obtained with \code{\link{get_protein_no_quant_mask }}. This can then be combined with this function
to replace protein level quantification values with NA where they were derived from too few quantification values
}
\examples{
mask <- get_protein_no_quant_mask(
  tmt_qf[["psms_filtered_forSum"]], min_features = 2)

masked <- mask_protein_level_quant(tmt_qf[["protein"]], mask)

sum(is.na(SummarizedExperiment::assay(masked)))
}
