% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_missingness.R
\name{mnar_global_score}
\alias{mnar_global_score}
\title{Deprecated: use global_condition_miss_score()}
\usage{
mnar_global_score(...)
}
\arguments{
\item{...}{arguments passed on to \code{\link{global_condition_miss_score}}}
}
\description{
\code{mnar_global_score()} has been renamed to
\code{\link{global_condition_miss_score}()} to better reflect that the score
measures missingness predictable from experimental condition, not MNAR in
the classical sense. All arguments are identical; this wrapper calls the new
function and will be removed in a future release.
}
\seealso{
\code{\link{global_condition_miss_score}}
}
