% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{plot_cor_samples}
\alias{plot_cor_samples}
\title{Plot the correlation between sample}
\usage{
plot_cor_samples(obj, i, order = "original", ...)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarizedExperiment you wish to plot}

\item{order}{\code{character} ordering method passed onto corrplot::corrplot}

\item{...}{addiional arguments passed onto corrplot::corrplot}
}
\value{
Returns a \emph{ggplot} object.
}
\description{
The correlation between feature quantifications in each sample can
be a useful QC step to assess whether the experiment has worked. This function plots
the correlation values in a heatmap. Samples are not clustered and are ordered
on the axes in the order they are in the SummarizedExperiment
}
\examples{

tmt_qf <- QFeatures::readQFeatures(assayData = psm_tmt_total,
  quantCols = 36:45,
  name = "psms_raw")

plot_cor_samples(tmt_qf, 'psms_raw')
}
