% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rt.R
\name{plot_rt_dist}
\alias{plot_rt_dist}
\title{Plot the retention time distribution}
\usage{
plot_rt_dist(obj, rt_col = "RT.in.min.by.Search.Engine.Sequest.HT")
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing peptide-level output from Proteome Discoverer.}

\item{rt_col}{\code{string}. Name of column with retention time}
}
\value{
Returns a \code{ggplot} with the RT vs delta PPM
}
\description{
It's useful to assess the retention time distribution as a quality control step.
The peptides should be spread across the retention times, without any clear 'gaps' or very clear
trend towards one end of the gradient.
}
\examples{
lfq_inf <- system.file("extdata", "lfq_dda_pd_PeptideGroups.txt",
                       package = "biomasslmb")

lfq_qf_raw <- QFeatures::readQFeatures(
  assayData = read.delim(lfq_inf),
  quantCols = grep("^Abundance", colnames(read.delim(lfq_inf))),
  name = "peptides")

plot_rt_dist(lfq_qf_raw[["peptides"]],
             rt_col = "PSM.RT.in.min.by.Search.Engine.CHIMERYS")
}
