% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rt.R
\name{plot_rt_vs_delta}
\alias{plot_rt_vs_delta}
\title{Plot the retention time vs delta precursor mass}
\usage{
plot_rt_vs_delta(
  obj,
  rt_col = "RT.in.min.by.Search.Engine.Sequest.HT",
  delta_ppm_col = "Delta.M.in.ppm.by.Search.Engine.Sequest.HT"
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing peptide-level output from Proteome Discoverer.}

\item{rt_col}{\code{string}. Name of column with retention time}

\item{delta_ppm_col}{\code{string}. Name of column with the Delta precursor mass}
}
\value{
Returns a \code{ggplot} with the RT vs delta PPM
}
\description{
It's useful to assess the relationship between retention time and delta precursor
mass as a quality control step. There should be little or no relationship
}
