% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot for differential abundance testing}
\usage{
plot_volcano(
  obj,
  lf_col = "logFC",
  p_col = "adj.P.Val",
  sig_col = NULL,
  title = NULL,
  xlab = "log2 Fold Change"
)
}
\arguments{
\item{obj}{\code{data.frame}. Statistical testing results}

\item{lf_col}{\code{string}. Column for log fold-change}

\item{p_col}{\code{string}. Column for p-value}

\item{sig_col}{\code{string}. Column for significance annotation (optional)}

\item{title}{\code{string}. Plot title}

\item{xlab}{\code{string}. Label for x axis}
}
\value{
Returns a \emph{ggplot} object.
}
\description{
Statistical testing results for 'omics' differential abundance
are frequently displayed in a 'volcano' plot, where x = log fold-change
and y = -log10(p-value). This function generates a volcano plot from a \code{data.frame}
containing the statistical testing results
}
\examples{
# plot_volcano takes a statistical testing results data.frame, such as
# the output of limma::topTable(), with logFC and adj.P.Val columns
results <- data.frame(
  logFC = rnorm(500),
  adj.P.Val = runif(500))

plot_volcano(results)
}
