% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_phospho.R
\docType{data}
\name{psm_tmt_phospho_total}
\alias{psm_tmt_phospho_total}
\title{PSM-level PD output for the total fraction matching \code{psm_tmt_phospho}}
\format{
An object of class \code{data.frame} with 17045 rows and 77 columns.
}
\usage{
psm_tmt_phospho_total
}
\description{
Proteome Discoverer PSM-level output for the total
(non-enriched) fraction of the experiment described in
\code{\link{psm_tmt_phospho}}, acquired from the same labelled pool and so
sharing its experimental design. Used to normalise the enriched fraction and
to express site abundance relative to protein abundance.
}
\keyword{datasets}
