% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_diann.R
\name{readDIANNFilterQJoin}
\alias{readDIANNFilterQJoin}
\title{Read in data from DIA-NN}
\usage{
readDIANNFilterQJoin(
  diann_report_infile,
  global_protein_q = 0.01,
  run_protein_q = 0.01,
  run_precursor_q = 0.01,
  return_sep_quant = FALSE
)
}
\arguments{
\item{diann_report_infile}{\code{character}. Filepath to DIA-NN report.tsv file}

\item{global_protein_q}{\code{numeric}. Threshold for global protein FDR}

\item{run_protein_q}{\code{numeric}. Threshold for run-specific protein FDR}

\item{run_precursor_q}{\code{numeric}. Threshold for run-specific precursor FDR}

\item{return_sep_quant}{\code{logical}. Return the individual sample-level quantification as well}
}
\description{
This function reads in the report.tsv file from DIA-NN, performs
filtering using the Q-Value columns and then joins together the quantification
from the individual samples
}
