% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{remove_redundant_go}
\alias{remove_redundant_go}
\title{Remove redundant GO terms}
\usage{
remove_redundant_go(
  obj,
  go_category_col = "category",
  p_value_col = "over_represented_pvalue"
)
}
\arguments{
\item{obj}{\code{data.frame} containing \code{goseq} results as generated by
\code{\link{get_enriched_go}} or \code{\link[goseq]{goseq}}.}

\item{go_category_col}{\code{character} column name for GO terms}

\item{p_value_col}{\code{character} column name for p-values}
}
\value{
Returns \code{obj} with redundant GO terms filtered out.
}
\description{
Given the output of \code{\link{get_enriched_go}} or
\code{\link[goseq]{goseq}}, remove redundant GO terms.
}
