% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_redundant_psms.R
\name{remove_redundant_psm_quant}
\alias{remove_redundant_psm_quant}
\title{Remove redundant Peptide Spectrum Matches}
\usage{
remove_redundant_psm_quant(
  obj,
  verbose = FALSE,
  master_protein_col = "Master.Protein.Accessions",
  validate = TRUE
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing PSM-level output from Proteome Discoverer.}

\item{verbose}{\code{boolean}. Don't output tallies of PSMs/proteins; Default=FALSE}

\item{master_protein_col}{\code{string}. Name of column containing master}

\item{validate}{\code{logical}. Check that sequence to protein assignments are consistent. Default=TRUE
proteins. (Only used when verbose=TRUE)}
}
\value{
Returns a \code{SummarisedExperiment} with the redundant quantification removed
}
\description{
When multiple search engines are used in Proteome Discoverer for TMT proteomics,
the PSM-level quantification contains duplicate rows for each search engine match
to the same spectrum. We can use the scan number columns
and quantitative data to identify the duplicate rows and exclude them.

The removal of duplicate quantification should be performed following initial
filtering to 1) remove rejected PSMs (see filter_TMT_PSMs) and 2)
retain only rank 1 matches (Filter on Rank and Search.Engine.Rank columns)
}
