% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{restrict_imputation}
\alias{restrict_imputation}
\title{Restrict imputed values to specific conditions}
\usage{
restrict_imputation(
  obj,
  i_unimputed,
  i_imputed,
  i_restricted_imputed,
  use_imputed_df,
  verbose = TRUE
)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i_unimputed}{\code{string}. Index for the SummarizedExperiment with data without imputation}

\item{i_imputed}{\code{string}. Index for the SummarizedExperiment with data with imputation}

\item{i_restricted_imputed}{\code{string}. Index for the output assay with restricted imputation}

\item{use_imputed_df}{\code{data.frame} with conditions where imputation should be used. see @details for more information}

\item{verbose}{\code{logical} Describe how many missing values in input and output assays}
}
\value{
Returns a \code{QFeatures} with restricted imputation with specified assay name
}
\description{
This function takes a \code{Qfeatures} object with unimputed data and
imputed data in separate assays and creates a new assay where the imputed values are
only used in specified circumstances. Note that this restriction occurs post imputation,
which may not be suitable for imputation methods where the imputed values are not independently derived
}
\details{
The \code{data.frame} provided to \code{use_imputed_df} should specify the conditions
under which imputation should be performed, with respect to the experimental
conditions and the number of missing values. \code{data.frame} must contain a column
called \code{n_finite} and all other columns must be columns in \code{colData(obj)}

For example, in an enrichment vs
control experiment, to only input in control samples where zero or one replicate
are quantified, this should be specified thusly

condition  n_finite
control  0
control  1
}
