% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_2plex.R
\docType{data}
\name{tmt_2plex_design}
\alias{tmt_2plex_design}
\title{Experimental design for \code{psm_tmt_2plex}}
\format{
An object of class \code{data.frame} with 29 rows and 7 columns.
}
\usage{
tmt_2plex_design
}
\description{
A \code{data.frame} with one row per labelled channel, giving the
Genotype, Time and Replicate for each sample in
\code{\link{psm_tmt_2plex}}. Channels that were not labelled in a given plex
are absent. \code{quantCols} gives the TMT tag and \code{runCol} the plex, as
required by \code{QFeatures::readQFeatures(runCol = 'Plex')}. The pooled bridge
channel present in both plexes has a \code{Genotype} of \code{Bridge}.
}
\keyword{datasets}
