% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_clock.R
\docType{data}
\name{tmt_clock_design}
\alias{tmt_clock_design}
\title{Experimental design for \code{psm_tmt_clock}}
\format{
An object of class \code{data.frame} with 12 rows and 4 columns.
}
\usage{
tmt_clock_design
}
\description{
A \code{data.frame} giving the Condition (\code{Control}/\code{Mutant}) and
Replicate for each sample in \code{\link{psm_tmt_clock}}, with rownames
matching the quantification column names.
}
\keyword{datasets}
