% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_per2_mq.R
\docType{data}
\name{tmt_per2_mq_design}
\alias{tmt_per2_mq_design}
\title{Experimental design for \code{psm_tmt_per2_mq}}
\format{
An object of class \code{data.frame} with 12 rows and 3 columns.
}
\usage{
tmt_per2_mq_design
}
\description{
A \code{data.frame} giving the Condition (\code{IP}/\code{Control}) and
Replicate for each sample in \code{\link{psm_tmt_per2_mq}}, with
rownames matching the quantification column names.
}
\keyword{datasets}
