% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_phospho.R
\docType{data}
\name{tmt_phospho_design}
\alias{tmt_phospho_design}
\title{Experimental design for \code{psm_tmt_phospho} and \code{psm_tmt_phospho_total}}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\usage{
tmt_phospho_design
}
\description{
A \code{data.frame} giving the Condition (\code{Control}/\code{Treated}),
Timepoint (\code{T1}/\code{T2}) and Replicate for each sample in
\code{\link{psm_tmt_phospho}} and \code{\link{psm_tmt_phospho_total}}, with
rownames matching the quantification column names.
}
\keyword{datasets}
