% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{tmt_qf}
\alias{tmt_qf}
\title{TMT data}
\format{
An object of class \code{QFeatures} of length 8.
}
\usage{
tmt_qf
}
\description{
\code{QFeatures} object holding the single-plex TMT dataset
(\code{psm_tmt_clock}) processed through PSM QC, filtering and summarisation to
protein level with \code{sum}, as produced in the
\verb{TMT workflow: PSM QC and protein summarisation} vignette.
}
\keyword{datasets}
