% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_per2_mq.R
\docType{data}
\name{tmt_qf_mq}
\alias{tmt_qf_mq}
\title{TMT data (MaxQuant input)}
\format{
An object of class \code{QFeatures} of length 5.
}
\usage{
tmt_qf_mq
}
\description{
\code{QFeatures} object containing the MaxQuant TMT IP-vs-control
data (\code{\link{psm_tmt_per2_mq}}) processed through PSM filtering and
\code{robustSummary} summarisation to protein level, as produced in Part A of
the \verb{Enrichment designs: IP, BioID and TurboID} vignette.
}
\keyword{datasets}
