% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_TMT_PSMs.R
\name{update_average_sn}
\alias{update_average_sn}
\title{Adds new feature describing the average reporter Signal/Noise ratio.}
\usage{
update_average_sn(obj, sn_col = "Average.Reporter.SN")
}
\arguments{
\item{obj}{\code{SummarizedExperiment}. Should contain PSMs-level TMT quantification}

\item{sn_col}{\code{string}. Name of output column containing the average signal:noise}
}
\value{
Returns an \code{MSnSet} with the average SN included as a new feature column.
}
\description{
PD column \code{Average.Reporter.SN} is NA when all tags have missing
values and imputes a value of zero for NA when there not all tags are missing.
So, with a 10-plex TMT experiment, if a single tag has a SN of 10 and all others
are NA, PD will report an average SN of 10/10 = 1.

This function adds an average reporter SN column which ignores missing
values. In the example above, the average SN value reported would be 10.
Where all values are NA, the average SN will remain NA.

The assumption is that the intensity values in the \code{exprs} matrix
are Signal/Noise ratios, as reported by PD by default
}
