% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_tm_toplogy.R
\name{add_tm_info}
\alias{add_tm_info}
\title{Add transmembrane domain details}
\usage{
add_tm_info(query_result)
}
\arguments{
\item{query_resul}{\code{data.frame} output of query_protein_tm_topology}
}
\value{
\code{data.frame} object.
}
\description{
Parse the Transmembrane column from the output of query_protein_tm_topology to add columns decscribing the transmembrane domains.
}
\examples{
uniprotIDs <- c('O76024', 'Q03135', 'Q96T23')
query_result <- query_protein_tm_topology(uniprotIDs)
query_result_parsed <- add_tm_info(query_result)
}
