% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_tm_toplogy.R
\name{add_topology_info}
\alias{add_topology_info}
\title{Add topology details}
\usage{
add_topology_info(result)
}
\arguments{
\item{query_resul}{\code{data.frame} output of query_protein_tm_topology}
}
\value{
\code{data.frame} object.
}
\description{
Parse the Topological.domain column from the output of query_protein_tm_topology to add columns decscribing the topology
}
\examples{
uniprotIDs <- c('O76024', 'Q03135', 'Q96T23')
query_result <- query_protein_tm_topology(uniprotIDs)
query_result_parsed <- add_topology_info(query_result)
}
