% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomasslmb-package.R
\docType{package}
\name{biomasslmb-package}
\alias{biomasslmb}
\alias{biomasslmb-package}
\title{biomasslmb: BioMass LMB Proteomics Data Analysis Utility Functions}
\description{
Utility functions for piping together quantitative proteomics data analysis. Designed for projects conducted with the Biological Mass Spectrometry and Proteomics Facility at the MRC Laboratory of Molecular Biology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lmb-mass-spec-compbio.github.io/biomasslmb/}
}

}
\author{
\strong{Maintainer}: Tom Smith \email{tsmith@mrc-lmb.cam.ac.uk} (\href{https://orcid.org/0000-0002-0697-8777}{ORCID})

}
\keyword{internal}
