% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{dia_qf}
\alias{dia_qf}
\title{LFQ-DIA data}
\format{
An object of class \code{QFeatures} of length 6.
}
\usage{
dia_qf
}
\description{
\code{QFeatures} object holding the LFQ-DIA plasma dataset
(\code{monkeypox_plasma_proteomes.parquet}) processed through precursor QC,
filtering and summarisation to protein level with \code{robustSummary}, as
produced in the
\verb{LFQ-DIA workflow: precursor QC and protein summarisation} vignette.
}
\keyword{datasets}
