% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{dia_qf}
\alias{dia_qf}
\title{DIA data}
\format{
An object of class \code{QFeatures} of length 5.
}
\usage{
dia_qf
}
\description{
Qfeature containing DIA data processed in the \verb{LFQ-DIA QC Peptide-level quantification and summarisation to protein-level abundance} vignette
}
\keyword{datasets}
