% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_protein_FDR.R
\name{filter_by_protein_fdr}
\alias{filter_by_protein_fdr}
\title{Filter to remove peptides from proteins failing the protein FDR threshold}
\usage{
filter_by_protein_fdr(
  obj,
  protein_fdr_filename,
  protein_col_peptide = "Master.Protein.Accessions",
  protein_col_protein = "Accession",
  protein_FDR_col = "Protein.FDR.Confidence.Combined",
  retain_proteins = NULL
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing peptide-level output from Proteome Discoverer.}

\item{protein_fdr_filename}{\code{string}. Filepath for protein-level output from Proteome Discoverer}

\item{protein_col_peptide}{\code{string}. Name of column containing master
proteins in the peptide-level \code{obj}}

\item{protein_col_protein}{\code{string}. Name of column containing master
proteins in the protein-level \code{protein_fdr_filename}}

\item{protein_FDR_col}{\code{string}. Name of column containing FDR information in
the protein-level \code{protein_fdr_filename}}

\item{retain_proteins}{\verb{character vector}. Vector of protein accessions to always retain,
even if they do not pass the FDR threshold. Default is NULL.}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the content of a summarised experiment
based on a separate file detailing the proteins passing the FDR confidence threshold

This function is used to filter peptides to retain only those from proteins
passing the FDR threshold
}
