% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw.R
\name{filter_features_diann}
\alias{filter_features_diann}
\title{Filter DIA-NN output}
\usage{
filter_features_diann(
  obj,
  master_protein_col = "Protein.Group",
  protein_col = "Protein.Ids",
  unique_master = TRUE,
  filter_contaminant = TRUE,
  contaminant_proteins = NULL,
  filter_associated_contaminant = TRUE,
  remove_no_quant = TRUE
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer.
Use \code{\link[QFeatures]{readQFeatures}} to read in .txt file}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{protein_col}{\code{string}. Name of column containing all protein
matches.}

\item{unique_master}{\code{logical}. Filter out features without a unique
master protein.}

\item{filter_contaminant}{\code{logical}. Filter out features which match a contaminant
protein.}

\item{contaminant_proteins}{\verb{character vector}. The protein IDs form the contaminant proteins}

\item{filter_associated_contaminant}{\code{logical}. Filter out features which
match a contaminant associated protein.}

\item{remove_no_quant}{\code{logical}. Remove features with no quantification}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the precursor .txt file from
DIA-NN, based on various criteria:
\enumerate{
\item Remove features without a master protein (Protein.Group column)
\item Remove features without a unique master protein
\item Remove features matching a contaminant protein
\item Remove features matching any protein associated with
a contaminant protein (see below)
\item Remove features without quantification values
}
}
\details{
\strong{Associated contaminant proteins} are proteins which have at least
one feature shared with a contaminant protein. It has been observed that the contaminant
fasta files often do not contain all possible contaminant proteins e.g. some features
can be assigned to a keratin which is not in the provided contaminant database.

In the example below, using \code{filter_associated_contaminant = TRUE} will filter out f2 and f3 in
addition to f1, regardless of the value in the Master.Protein.Accession column.

\if{html}{\out{<div class="sourceCode">}}\preformatted{feature  Protein.Accessions         Master.Protein.Accessions
f1       protein1, protein2, contaminant,  protein1,
f2       protein1, protein3         protein3,
f3       protein2                   protein2
}\if{html}{\out{</div>}}
}
