% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw.R
\name{filter_features_pd_dda}
\alias{filter_features_pd_dda}
\title{Filter Proteome Discoverer DDA output}
\usage{
filter_features_pd_dda(
  obj,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  filter_contaminant = TRUE,
  contaminant_proteins = NULL,
  crap_proteins = NULL,
  filter_associated_contaminant = TRUE,
  remove_no_quant = TRUE,
  cont_string = "Cont_"
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer.
Use \code{\link[QFeatures]{readQFeatures}} to read in .txt file}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{protein_col}{\code{string}. Name of column containing all protein
matches.}

\item{unique_master}{\code{logical}. Filter out features without a unique
master protein.}

\item{filter_contaminant}{\code{logical}. Filter out features which match a contaminant
protein.}

\item{contaminant_proteins}{\verb{character vector}. The protein IDs form the contaminant proteins}

\item{crap_proteins}{\verb{character vector}. Same as contaminant_proteins. Available for backwards compatibility. Default is NULL. If both contaminant_proteins and crap_proteins are set, an error is thrown.}

\item{filter_associated_contaminant}{\code{logical}. Filter out features which
match a contaminant associated protein.}

\item{remove_no_quant}{\code{logical}. Remove features with no quantification}

\item{cont_string}{\code{string}. string to search for contaminants}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the output .txt files (peptide groups or PSMs) from
Proteome Discoverer for DDA, based on various criteria:
\enumerate{
\item Remove features without a master protein
\item Remove features without a unique master protein  (i.e.
Number.of.Protein.Groups == 1)
\item Remove features matching a contaminant protein
\item Remove features matching any protein associated with
a contaminant protein (see below)
\item Remove features without quantification values
}
}
\details{
\strong{Associated contaminant proteins} are proteins which have at least
one feature shared with a contaminant protein. It has been observed that the contaminant
fasta files often do not contain all possible contaminant proteins e.g. some features
can be assigned to a keratin which is not in the provided contaminant database.

In the example below, using \code{filter_associated_contaminant = TRUE} will filter out f2 and f3 in
addition to f1, regardless of the value in the Master.Protein.Accession column.

\if{html}{\out{<div class="sourceCode">}}\preformatted{feature  Protein.Accessions         Master.Protein.Accessions
f1       protein1, protein2, contaminant,  protein1,
f2       protein1, protein3         protein3,
f3       protein2                   protein2
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{

#### PSMs.txt example ####
# load PD PSMs.txt output
tmt_qf <- readQFeatures(assayData = psm_tmt_total,
 quantCols = 36:45,
 name = "psms_raw")

# extract the UniProt accessions from the contaminant FASTA headers
contaminant_accessions <- get_crap_fasta_accessions(contaminant_fasta_inf)

# filter the PSMs
psm2 <- filter_features_pd_dda(
  obj = tmt_qf[['psms_raw']],
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  unique_master = TRUE,
  TMT = TRUE,
  filter_contaminant = TRUE,
  contaminant_proteins = contaminant_accessions,
  filter_associated_contaminant = TRUE
)


}
}
