% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_raw.R
\name{filter_features_sn}
\alias{filter_features_sn}
\title{Filter Spectronaut output}
\usage{
filter_features_sn(
  obj,
  master_protein_col = "PG.ProteinAccessions",
  unique_master = TRUE,
  filter_contaminant = TRUE,
  contaminant_proteins = NULL,
  remove_no_quant = TRUE,
  cont_string = "Cont_"
)
}
\arguments{
\item{obj}{\code{SummarisedExperiment} containing output from Proteome Discoverer.
Use \code{\link[QFeatures]{readQFeatures}} to read in .txt file}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{unique_master}{\code{logical}. Filter out features without a unique
master protein.}

\item{filter_contaminant}{\code{logical}. Filter out features which match a contaminant
protein.}

\item{contaminant_proteins}{\verb{character vector}. The protein IDs form the contaminant proteins}

\item{remove_no_quant}{\code{logical}. Remove features with no quantification}

\item{cont_string}{\code{string}. string to search for contaminants}
}
\value{
Returns a \code{SummarisedExperiment} with the filtered Proteome Discoverer output.
}
\description{
This function filters the precursor .txt file from
Spectronaut, based on various criteria:
\enumerate{
\item Remove features without a master protein (PG.ProteinAccessions column)
\item Remove features without a unique master protein
\item Remove features matching a contaminant protein
\item Remove features without quantification values
}
}
