% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho.R, R/ptm.R
\name{filter_maxquant_ptm}
\alias{filter_maxquant_ptm}
\title{Make a FASTA using UniProt accessions}
\usage{
filter_maxquant_ptm(obj, i, ptm_prob_col = "Phospho..STY..Probabilities")

filter_maxquant_ptm(obj, i, ptm_prob_col = "Phospho..STY..Probabilities")
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarizedExperiment with data without imputation}

\item{ptm_prob_col}{\code{character} column name for PTMs.}

\item{accessions}{\verb{character vector}, the UniProt accessions to use}

\item{file}{\code{character}, filepath to save the fasta to e.g. \code{"crap.fasta"}}

\item{is_crap}{\code{logical}, Is the output going to be a cRAP database? If \code{TRUE}
cRAP001, cRAP002, etc. is appended to the sequence headers in the FASTA file.
Default is \code{FALSE}}

\item{overwrite}{\code{logical}, if the FASTA file already exists should it be
overwritten? Default is \code{FALSE}}

\item{verbose}{\code{logical}, should the function send messages to the console?
Default is \code{TRUE}}
}
\value{
Returns a FASTA file saved to disk at the specified file path.

Returns a \code{SummarizedExperiment} where the value for the PTM probability column is not empty
}
\description{
Given a vector of UniProt accessions, this function will:
\enumerate{
\item Download the sequences
\item Print the current UniProt release (you should put this in the FASTA file name)
\item (Optional) add cRAP numbers to the FASTA headers for each sequence
\item Save the sequences in a FASTA file
}

This function takes a \code{Qfeatures} object with the evidence.txt
output from MaxQuant and filters it to retain peptides with a given PTM.
}
