% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{lfq_qf}
\alias{lfq_qf}
\title{LFQ-DDA data}
\format{
An object of class \code{QFeatures} of length 6.
}
\usage{
lfq_qf
}
\description{
\code{QFeatures} object holding the LFQ-DDA whole-proteome dataset
(\code{lfq_dda_pd_PeptideGroups.txt}) processed through peptide QC, filtering and
summarisation to protein level with \code{robustSummary}, as produced in the
\verb{LFQ-DDA workflow: peptide QC and protein summarisation} vignette.
}
\keyword{datasets}
