% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_missingness.R
\name{mnar_score}
\alias{mnar_score}
\title{Deprecated: use condition_miss_score()}
\usage{
mnar_score(...)
}
\description{
\code{mnar_score()} has been renamed to \code{\link{condition_miss_score}()}
to better reflect that the score measures missingness predictable from
experimental condition, not MNAR in the classical sense. All arguments are
identical; this wrapper calls the new function and will be removed in a
future release.
}
\seealso{
\code{\link{condition_miss_score}}
}
