% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{plot_go_terms_upset}
\alias{plot_go_terms_upset}
\title{Plot selected GO terms}
\usage{
plot_go_terms_upset(
  goi,
  foi,
  gene2cat,
  term_col = "TERM",
  gene_col = "UNIPROTKB"
)
}
\arguments{
\item{goi}{\code{character} GO terms of interest}

\item{foi}{\code{character} features (genes/proteins etc) of interest}

\item{gene2cat}{\code{data.frame} with 2 columns containing the mapping between
genes (usually UniProt accessions, Ensembl gene IDs or similar) and GO terms.}

\item{term_col}{\code{character} column name for GO term description.}

\item{gene_col}{\code{character} column name for genes (or protein etc).}
}
\value{
Returns a \code{ggplot} object.
}
\description{
This function plots a set of GO terms of interest.
be run after \code{\link{get_enriched_go}} and \code{\link{estimate_go_overrep}}.
To avoid plotting too many terms, you may wish to use
\code{\link{remove_redundant_go}} first too. A warning is shown if you try
and plot more than 100 GO terms.
}
