% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{plot_missing_upset}
\alias{plot_missing_upset}
\title{A ggplot2 theme for the package}
\usage{
plot_missing_upset(obj, i)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarisedExperiment you wish to plot}
}
\value{
Returns a \emph{ggplot} object.
}
\description{
A standard theme to use for plots generated by this package.
}
\examples{
set.seed(11)
library(ggplot2)

df <- diamonds[sample(nrow(diamonds), 1000), ]

tmt_qf <- QFeatures::readQFeatures(assayData = psm_tmt_total,
  quantCols = 36:45,
  name = "psms_raw")






}
