% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{plot_missing_upset}
\alias{plot_missing_upset}
\title{Plot the most common missing value patterns}
\usage{
plot_missing_upset(obj, i)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarisedExperiment you wish to plot}
}
\value{
Returns a \emph{ggplot} object.
}
\description{
The patterns in missing values can be informative with respect to
whether the experiment has worked, or if particular samples are outliers. This
function uses an 'upset' plot to show the top 50 most common missing value patterns
across the samples
}
\examples{
set.seed(11)
library(ggplot2)

df <- diamonds[sample(nrow(diamonds), 1000), ]

tmt_qf <- QFeatures::readQFeatures(assayData = psm_tmt_total,
  quantCols = 36:45,
  name = "psms_raw")






}
