% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Create a Principle Component plot from the feature quantification}
\usage{
plot_pca(obj, i, colour_by = NULL, shape_by = NULL, x = 1, y = 2)
}
\arguments{
\item{obj}{\code{QFeatures}. Proteomics dataset}

\item{i}{\code{string}. Index for the SummarizedExperiment you wish to plot}

\item{colour_by}{\code{string}. ColData column to colour points by}

\item{shape_by}{\code{string}. ColData column to shape points by}

\item{x}{\code{numeric}. Principle component to plot on x-axis}

\item{y}{\code{numeric}. Principle component to plot on x-axis}
}
\value{
Returns a \emph{ggplot} object.
}
\description{
A PCA visualisation of feature quantifications in each sample can
allow one to see how the experimental conditions relate to the sources of variance (principle components).
This function plots a PCA, with the option to colour and/or shape the points by experimental conditions.
The percentage values indicated on the axes are the variance explained by the PCs.
}
