% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_protein_assay.R
\name{plot_protein_assays}
\alias{plot_protein_assays}
\title{Plot quantification at multiple levels}
\usage{
plot_protein_assays(
  obj,
  poi,
  experiments_to_plot = NULL,
  protein_id_col = "Master.Protein.Accessions",
  label_col = "Master.Protein.Accessions",
  rename_labels = NULL,
  log2transform_cols = "",
  norm_quant = FALSE,
  add_mean_summary = FALSE,
  colour_assays = FALSE,
  alpha_assays = TRUE
)
}
\arguments{
\item{obj}{\code{QFeatures} object}

\item{poi}{\code{list}. Proteins of interest}

\item{experiments_to_plot}{\code{list}. experiments (assays) to plot. Defaults to all assays}

\item{protein_id_col}{\code{string}. Column with the protein ids to search for \code{poi} values}

\item{label_col}{\code{string}. Column with labels to use for proteins in plot}

\item{rename_labels}{\verb{named list}. Mapping from labels to renamed labels}

\item{log2transform_cols}{\code{list}. Assays which need to be log2-transforms (all values should ultimately be transformed)}

\item{norm_quant}{\code{logical}. Should the quantifications be normalised to the fold-change vs mean abundance}

\item{add_mean_summary}{\code{logical}. Add a line summarising the mean value over all features in the assay (Not recommended if norm_quant=FALSE)}

\item{colour_assays}{\code{logical}. Column each assay}

\item{alpha_assays}{\code{logical}. Set sensible alpha value for each assay}
}
\value{
\code{ggplot} object.
}
\description{
For the exploration of specific proteins, it's often useful to visualise the quantification at multiple levels,
e.g PSM, peptide, filtered peptides and protein.
This function plots the quantitative data at multiple levels (assays) for a given list of proteins of interest.
}
