% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_quant}
\alias{plot_quant}
\title{Plot distributions for feature intensities per sample.}
\usage{
plot_quant(
  obj,
  method = c("box", "density", "histogram"),
  log2transform = FALSE,
  facet_by_sample = FALSE
)
}
\arguments{
\item{obj}{\code{SummarizedExperiment}.}

\item{method}{\code{string}. Plot style. Choice of box, density or histogram plot.}

\item{log2transform}{\code{logical}. Should feature quantifications be log-transformed?}

\item{facet_by_sample}{\code{logical}. Facet the plot by sample.}
}
\value{
\code{ggplot} object.
}
\description{
Given a \code{SummarizedExperiment}, return a plot
of the feature quantifications per sample.
}
