% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_detection.R
\name{plot_samples_present}
\alias{plot_samples_present}
\title{Plots the number of samples each feature was detected in for each experiment in a Qfeatures object}
\usage{
plot_samples_present(samples_present, rowvars, breaks = NULL)
}
\arguments{
\item{samples_present}{\code{data.frame} output from \code{get_samples_present}}

\item{rowvars}{\verb{character vector} row variables on which the tallying was grouped by}

\item{breaks}{\verb{numeric vector} breaks for the 'Samples' fill scale. If NULL (default), breaks are determined automatically}
}
\value{
\code{ggplot} object.
}
\description{
When exploring the proteome coverage, it is instructive to consider how many samples a protein
was present in for each level of filtering. This function plots the output of \code{get_samples_present}
}
\examples{
rename_cols <- c("All PSMs" = "psms_raw", "Protein" = "protein")

samples_present <- get_samples_present(
  tmt_qf, rowVars = "Master.Protein.Accessions", rename_cols = rename_cols)

plot_samples_present(samples_present, rowvars = "Master.Protein.Accessions")
}
