% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_total.R, R/tmt_qf.R
\docType{data}
\name{psm_tmt_total}
\alias{psm_tmt_total}
\title{PSM-level PD output for total proteome TMT10-plex data}
\format{
An object of class \code{data.frame} with 5000 rows and 50 columns.

An object of class \code{data.frame} with 5000 rows and 50 columns.
}
\usage{
psm_tmt_total

psm_tmt_total
}
\description{
x

Qfeature containing TMT data processed in the TMT QC PSM-level quantification and summarisation to protein-level abundance vignette
}
\keyword{datasets}
