% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_tm_toplogy.R
\name{query_protein_tm_topology}
\alias{query_protein_tm_topology}
\title{Query UniProt to determine protein transmembrane-domains and topology}
\usage{
query_protein_tm_topology(uniprotIDs, verbosity = 0)
}
\arguments{
\item{uniprotIDs}{\verb{character vector} Uniprot IDs}

\item{verbosity}{\code{integer} Verbosity level for uniprotREST::uniprot_map}
}
\value{
\code{data.frame} object.
}
\description{
Given a set of UniProt IDs, this function returns the information regarding transmembrane domains and topology.
}
\examples{
uniprotIDs <- c('O76024', 'Q03135', 'Q96T23')
query_protein_tm_topology(uniprotIDs)
}
