% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls.r
\name{tls}
\alias{tls}
\title{Total least squares regression (for use with geom_smooth)}
\usage{
tls(formula, data, ...)
}
\arguments{
\item{formula}{Formula of the form y ~ x.}

\item{data}{A data frame containing columns for x and y.}

\item{...}{Ignored.}
}
\description{
Total least squares regression (for use with geom_smooth)
}
