% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qf_data.R
\docType{data}
\name{tmt_qf}
\alias{tmt_qf}
\title{TMT data}
\format{
An object of class \code{QFeatures} of length 7.
}
\usage{
tmt_qf
}
\description{
Qfeature containing TMT data processed in the \verb{TMT QC PSM-level quantification and summarisation to protein-level abundance} vignette
}
\keyword{datasets}
